% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_extrafont.R
\name{check_extrafont}
\alias{check_extrafont}
\title{Function to make sure necessary fonts are imported from system}
\usage{
check_extrafont()
}
\description{
function to make sure that extrafont package is installed properly and fonts are imported and loaded to be used in the pacakge
}
\details{
function to make sure that extrafont package is installed properly and fonts are imported and loaded to be used in the pacakge

To use custom fonts, please install the \code{extrafont} package and run \code{extrafont::font_import()} and \code{extrafont::loadfonts()}.
}
