% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation_multi()}
\alias{correlation_multi()}
\title{Run the Shiny App for Correlation Analysis of multiple variables}
\usage{
correlation_multi()
}
\value{
This function runs a local instance of the Shiny app in your default web
browser. The app interface allows users to upload data, select analysis
methods, and download outputs.
}
\description{
App allows user to import data through excel/.csv files or through clipboard
and select the correlation method and download the results with few customization options
}
\details{
This Shiny app is part of the \code{visvaR} package and is designed for
correlation analysis and user can download the report in word format also has
option to download the correlation plot as .png file
}
\author{
Ramesh Ramasamy

Mathiyarsai Kulandaivadivel

Tamilselvan Arumugam
}
