% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_rbd.R
\name{twoway_anova_rbd}
\alias{twoway_anova_rbd}
\title{Run the Shiny App for performing one-way ANOVA on data from a Randomized block design
App allows user to import data through excel/.csv files or through clipboard
and user can select the post-hoc test method and download the report which contains anova results and plots}
\usage{
twoway_rbd()
}
\value{
This function runs a local instance of the Shiny app in your default web
browser. The app interface allows users to upload data, select analysis
method, and download outputs.
}
\description{
Run the Shiny App for performing one-way ANOVA on data from a Randomized block design
App allows user to import data through excel/.csv files or through clipboard
and user can select the post-hoc test method and download the report which contains anova results and plots
}
\details{
This Shiny app is part of the \code{visvaR} package and is designed for
analysis of variance randomized block design (two factor) and user can download the report in word format.
}
\author{
Ramesh Ramasamy
}
