% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_rbd.R
\name{twoway_rbd}
\alias{twoway_rbd}
\alias{twoway_rbd()}
\title{Run the Shiny App for performing one-way ANOVA on data from a Randomized block design}
\usage{
twoway_rbd()
}
\value{
This function runs a local instance of the Shiny app in your default web
browser. The app interface allows users to upload data, select analysis
method, and download outputs.
}
\description{
App allows user to import data through excel/.csv files or through clipboard
and user can select the post-hoc test method and download the report which contains anova results and plots
}
\details{
This Shiny app is part of the \code{visvaR} package and is designed for
analysis of variance randomized block design (two factor) and user can download the report in word format.
}
\references{
Fisher, R. A. (1925). Statistical Methods for Research Workers. Oliver and Boyd, Edinburgh.
Scheffe, H. (1959). The Analysis of Variance. John Wiley & Sons, New York.
R Core Team (2024). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/
}
\author{
Ramesh Ramasamy

Mathiyarsai Kulandaivadivel

Tamilselvan Arumugam
}
